refactor namespaces for entities and finalize prototypical properties
collection, blocks and properties got their own namespace (within entities) for improved readability/structure
