+++ src/Entities/Collections/BlockCollection.php
@@ -1,20 +1,21 @@
 <?php
 
-namespace FiveamCode\LaravelNotionApi\Entities;
+namespace FiveamCode\LaravelNotionApi\Entities\Collections;
 
+use FiveamCode\LaravelNotionApi\Entities\Blocks\Block;
 use FiveamCode\LaravelNotionApi\Exceptions\WrapperException;
 use FiveamCode\LaravelNotionApi\Notion;
 use Illuminate\Support\Arr;
 use Illuminate\Support\Collection;
 
 
-class DatabaseCollection extends EntityCollection
+class BlockCollection extends EntityCollection
 {
     protected function collectChildren()
     {
         $this->collection = new Collection();
-        foreach ($this->rawResults as $databaseChild) {
-            $this->collection->add(new Database($databaseChild));
+        foreach ($this->rawResults as $blockChild) {
+            $this->collection->add(new Block($blockChild));
         }
     }
 }
