+++ src/Entities/Collections/DatabaseCollection.php
@@ -1,20 +1,21 @@
 <?php
 
-namespace FiveamCode\LaravelNotionApi\Entities;
+namespace FiveamCode\LaravelNotionApi\Entities\Collections;
 
+use FiveamCode\LaravelNotionApi\Entities\Database;
 use FiveamCode\LaravelNotionApi\Exceptions\WrapperException;
 use FiveamCode\LaravelNotionApi\Notion;
 use Illuminate\Support\Arr;
 use Illuminate\Support\Collection;
 
 
-class UserCollection extends EntityCollection
+class DatabaseCollection extends EntityCollection
 {
     protected function collectChildren()
     {
         $this->collection = new Collection();
-        foreach ($this->rawResults as $userChild) {
-            $this->collection->add(new User($userChild));
+        foreach ($this->rawResults as $databaseChild) {
+            $this->collection->add(new Database($databaseChild));
         }
     }
 }
