+++ src/Entities/Collections/PageCollection.php
@@ -1,20 +1,21 @@
 <?php
 
-namespace FiveamCode\LaravelNotionApi\Entities;
+namespace FiveamCode\LaravelNotionApi\Entities\Collections;
 
+use FiveamCode\LaravelNotionApi\Entities\Page;
 use FiveamCode\LaravelNotionApi\Exceptions\WrapperException;
 use FiveamCode\LaravelNotionApi\Notion;
 use Illuminate\Support\Arr;
 use Illuminate\Support\Collection;
 
 
-class BlockCollection extends EntityCollection
+class PageCollection extends EntityCollection
 {
     protected function collectChildren()
     {
         $this->collection = new Collection();
-        foreach ($this->rawResults as $blockChild) {
-            $this->collection->add(new Block($blockChild));
+        foreach ($this->rawResults as $pageChild) {
+            $this->collection->add(new Page($pageChild));
         }
     }
 }
