+++ src/Entities/Collections/UserCollection.php
@@ -1,20 +1,21 @@
 <?php
 
-namespace FiveamCode\LaravelNotionApi\Entities;
+namespace FiveamCode\LaravelNotionApi\Entities\Collections;
 
+use FiveamCode\LaravelNotionApi\Entities\User;
 use FiveamCode\LaravelNotionApi\Exceptions\WrapperException;
 use FiveamCode\LaravelNotionApi\Notion;
 use Illuminate\Support\Arr;
 use Illuminate\Support\Collection;
 
 
-class PageCollection extends EntityCollection
+class UserCollection extends EntityCollection
 {
     protected function collectChildren()
     {
         $this->collection = new Collection();
-        foreach ($this->rawResults as $pageChild) {
-            $this->collection->add(new Page($pageChild));
+        foreach ($this->rawResults as $userChild) {
+            $this->collection->add(new User($userChild));
         }
     }
 }
